(2.0): Restructure the library to use interfaces instead of classes
Classes aren't the best option for this library because they add friction
between the Slack API docs and the use of the library. It would require
learning both the Slack API and the library, so it's not wanted by anyone.
Instead, we are now using interfaces to provide strong intellisense and
documentation to the consumer when designing UI with BlockKit or sending
processing/sending rich messages using the Slack API.

Still awaiting for a sandbox environment to write tests atm.